Chrismastree
This Code generates a chrismastree in a height from 1-5, which optional can have a star on its top. To run the code, you first have to decide which values you type in at the top of the code(int height=(1-5);bool star=(true or false)).After you copied the code in Visual Studio and after typing in the values, just activate the "Run" at the top and you will get a window with your generated chrismastree.